新增设置项【所有任务完成计算后使用颜色来标记相同的哈希值】并实现功能，以及重新分类一些设置项

+++ HashCalculator/Native/Methods.cs
@@ -184,7 +184,7 @@ namespace HashCalculator
         /// https://learn.microsoft.com/en-us/windows/win32/api/shlwapi/nf-shlwapi-colorhlstorgb
         /// </summary>
         [DllImport("shlwapi.dll")]
-        public static extern uint ColorHLSToRGB(int wHue, int wLuminance, int wSaturation);
+        public static extern uint ColorHLSToRGB(ushort wHue, ushort wLuminance, ushort wSaturation);
     }
 
     internal static class SHELL32
